renaming convetion to the other addon
