+++ springboot/addons/messaging/implementation/src/main/java/org/kie/kogito/addon/messaging/common/EventUnmarshallerProducer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.addon.cloudevents.spring;
+package org.kie.kogito.addon.messaging.common;
 
 import org.kie.kogito.event.CloudEventUnmarshallerFactory;
 import org.kie.kogito.event.EventUnmarshaller;
