+++ springboot/addons/messaging/implementation/src/main/java/org/kie/kogito/addon/messaging/common/SpringMessageConsumer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.addon.cloudevents.spring;
+package org.kie.kogito.addon.messaging.common;
 
 import java.util.Collections;
 import java.util.concurrent.ExecutorService;
